+++ src/lava.h
@@ -12,7 +12,7 @@ void destroy_lava(lava_t *lava);
 int lava_render(const lava_t *lava,
                 SDL_Renderer *renderer,
                 const camera_t *camera);
-int lava_update(lava_t *lava, Uint32 delta_time);
+int lava_update(lava_t *lava, float delta_time);
 
 int lava_overlaps_rect(const lava_t *lava,
                        rect_t rect);
