+++ src/player.h
@@ -22,7 +22,7 @@ int player_sound(const player_t * player,
                  sound_medium_t * sound_medium);
 void player_update(player_t * player,
                    const platforms_t *platforms,
-                   Uint32 delta_time);
+                   float delta_time);
 
 void player_move_left(player_t *player);
 void player_move_right(player_t *player);
